Updated unit tests to match new behaviour
The middlwarePriority allows us to hide the fact a given file or file
category exists to begin with, instead of returning 404s on some
scenarios.

This is a lot safer than giving 403s on some paths and 404s on others,
as you don't provide any information instead of minimal information.