Removed BotMan and bot commands
It's poorly maintained, and pull requests remain open for a very long
time, which causes incompatibility issues with Laravel 7.

This would mean that a side-project prevents us from switching to a
better system with Laravel 7 with benefits like the Guzzle wrapper
(less complexity) and Laravel Sanctum (better API auth).

The site should be just the site. I've gotten sidetracked here.