Fixed sticky activities
Some activities were postponed, but later on cancelled. These are still
showing up, which is wrong. This adds a stricter check to the
`postponed_at` property, allowing it only if it's not used on a
cancelled or rescheduled activity

I switched to using the `Date` facade, instead of `now()` functions.